requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.7.1
PyPDF2==3.0.1
chromadb==0.6.3
sentence-transformers==3.3.1
pytest==8.3.4
httpx==0.28.1
